Доработки для x64. Перенос последних правок сообщества из прошлого снегопата.

+++ v8api/command.v
@@ -7,7 +7,7 @@
 	Guid group
 	uint num
 :meths
-	void ctor(const Guid&in group, int num)
+	void ctor(const Guid&in group, size_t num)
 	{
 		obj.group = group;
 		obj.num = num;
@@ -20,7 +20,7 @@
 	int       param
 	int_ptr	  object
 :meths
-	void ctor(const CommandID&in _id, int p)
+	void ctor(const CommandID&in _id, size_t p)
 	{
 		obj.id = _id;
 		obj.param = p;
